+++ src/main/java/services/DroneApi/DroneApiService.java
@@ -2,7 +2,7 @@ package main.java.services.DroneApi;
 
 import com.google.gson.Gson;
 import main.java.exceptions.DroneApiException;
-import main.java.services.DroneApi.DTOs.*;
+import main.java.services.DroneApi.dtos.*;
 
 import java.net.URI;
 import java.net.http.HttpClient;
