+++ src/main/java/services/DroneApi/IDroneApiService.java
@@ -1,8 +1,8 @@
 package main.java.services.DroneApi;
 
-import main.java.services.DroneApi.DTOs.Drone;
-import main.java.services.DroneApi.DTOs.DroneDynamics;
-import main.java.services.DroneApi.DTOs.DroneType;
+import main.java.services.DroneApi.dtos.Drone;
+import main.java.services.DroneApi.dtos.DroneDynamics;
+import main.java.services.DroneApi.dtos.DroneType;
 import main.java.exceptions.DroneApiException;
 
 import java.util.ArrayList;
