+++ src/main/java/services/DroneApi/dtos/DroneDynamicsResponse.java
@@ -1,4 +1,4 @@
-package main.java.services.DroneApi.DTOs;
+package main.java.services.DroneApi.dtos;
 
 import java.util.ArrayList;
 
@@ -6,7 +6,7 @@ import java.util.ArrayList;
  * DTO (Data Transfer Object) for Api Response
  * You probably shouldn't use this in your own code
  */
-public class DroneTypesResponse {
+public class DroneDynamicsResponse {
 
     /**
      * Number of total results available on the API
@@ -24,6 +24,6 @@ public class DroneTypesResponse {
     public String previous = "";
 
     /// The Results
-    public ArrayList<DroneType> results = new ArrayList<>();
+    public ArrayList<DroneDynamics> results = new ArrayList<DroneDynamics>();
 
 }
